requirements.txt
matplotlib
numpy==1.25.2
opencv_python
pandas
Pillow==10.0.0
PyYAML==6.0.1
scipy
seaborn==0.11.2
tqdm==4.64.1
